orchestrator/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
docker==7.1.0
requests==2.32.3
APScheduler==3.10.4
PyJWT==2.8.0
Werkzeug==3.0.3
